Fix gpr_log format mismatches in c++ tests

+++ test/cpp/end2end/async_end2end_test.cc
@@ -231,7 +231,8 @@ class TestScenario {
   void Log() const {
     gpr_log(GPR_INFO,
             "Scenario: disable_blocking %d, credentials %s, message size %d",
-            disable_blocking, credentials_type.c_str(), message_content.size());
+            disable_blocking, credentials_type.c_str(),
+            (int)(message_content.size()));
   }
   bool disable_blocking;
   const grpc::string credentials_type;
